+++ Gruntfile.js
@@ -1,10 +1,44 @@
 module.exports = function (grunt) {
   'use strict';
 
+  var liveReloadPort = 35729;
+
   require('load-grunt-tasks')(grunt);
 
   grunt.initConfig({
     pkg: grunt.file.readJSON('bower.json'),
+    // Files to watch for changes in order to make the browser reload
+    watch: {
+      options: {
+        livereload: liveReloadPort
+      },
+      livereload: {
+        files: [
+          'index.html',
+          'src/**/*.js'
+        ]
+      },
+    },
+
+    // Fires up a simple connect server - useful for development
+    connect: {
+      options: {
+        port: 9999,
+        hostname: 'localhost',
+        livereload: liveReloadPort
+      },
+      livereload: {
+        options: {
+          open: true,
+          base: [
+            // Directories to serve static files from
+            '.',
+            'app',
+            'src',
+          ]
+        }
+      }
+    },
     uglify: {
       options: {
         preserveComments: 'some'
@@ -34,11 +68,7 @@ module.exports = function (grunt) {
         autoWatch: true
       }
     },
-    changelog: {
-      options: {
-        github: 'passy/angular-masonry'
-      }
-    },
+    changelog: {},
     ngmin: {
       dist: {
         src: '<%= pkg.name %>.js',
@@ -48,5 +78,6 @@ module.exports = function (grunt) {
   });
 
   grunt.registerTask('default', ['concat', 'ngmin', 'uglify']);
+  grunt.registerTask('server', ['default', 'connect:livereload', 'watch']);
   grunt.registerTask('test', ['karma:dist']);
 };