+++ angular-masonry.js
@@ -1,3 +1,8 @@
+/*!
+ * angular-masonry 0.8.1
+ * Pascal Hartig, weluse GmbH, http://weluse.de/
+ * License: MIT
+ */
 (function () {
   'use strict';
   angular.module('wu.masonry', []).controller('MasonryCtrl', [
@@ -11,6 +16,7 @@
       var self = this;
       var timeout = null;
       this.preserveOrder = false;
+      this.loadImages = true;
       this.scheduleMasonryOnce = function scheduleMasonryOnce() {
         var args = arguments;
         var found = schedule.filter(function filterFn(item) {
@@ -20,6 +26,8 @@
           this.scheduleMasonry.apply(null, arguments);
         }
       };
+      // Make sure it's only executed once within a reasonable time-frame in
+      // case multiple elements are removed or added at once.
       this.scheduleMasonry = function scheduleMasonry() {
         if (timeout) {
           $timeout.cancel(timeout);
@@ -47,15 +55,23 @@
             $element.masonry('resize');
           }
           if (bricks[id] === undefined) {
+            // Keep track of added elements.
             bricks[id] = true;
             defaultLoaded(element);
             $element.masonry('appended', element, true);
           }
         }
         function _layout() {
+          // I wanted to make this dynamic but ran into huuuge memory leaks
+          // that I couldn't fix. If you know how to dynamically add a
+          // callback so one could say <masonry loaded="callback($element)">
+          // please submit a pull request!
           self.scheduleMasonryOnce('layout');
         }
-        if (self.preserveOrder) {
+        if (!self.loadImages) {
+          _append();
+          _layout();
+        } else if (self.preserveOrder) {
           _append();
           element.imagesLoaded(_layout);
         } else {
@@ -76,6 +92,7 @@
       this.destroy = function destroy() {
         destroyed = true;
         if ($element.data('masonry')) {
+          // Gently uninitialize if still present
           $element.masonry('destroy');
         }
         $scope.$emit('masonry.destroyed');
@@ -95,9 +112,11 @@
           var attrOptions = scope.$eval(attrs.masonry || attrs.masonryOptions);
           var options = angular.extend({
               itemSelector: attrs.itemSelector || '.masonry-brick',
-              columnWidth: parseInt(attrs.columnWidth, 10)
+              columnWidth: parseInt(attrs.columnWidth, 10) || attrs.columnWidth
             }, attrOptions || {});
           element.masonry(options);
+          var loadImages = scope.$eval(attrs.loadImages);
+          ctrl.loadImages = loadImages !== false;
           var preserveOrder = scope.$eval(attrs.preserveOrder);
           ctrl.preserveOrder = preserveOrder !== false && attrs.preserveOrder !== undefined;
           scope.$emit('masonry.created', element);
@@ -118,7 +137,8 @@
             ctrl.removeBrick(id, element);
           });
           scope.$on('masonry.reload', function () {
-            ctrl.reload();
+            ctrl.scheduleMasonryOnce('reloadItems');
+            ctrl.scheduleMasonryOnce('layout');
           });
           scope.$watch('$index', function () {
             if (index !== undefined && index !== scope.$index) {