+++ karma.conf.js
@@ -5,7 +5,7 @@ module.exports = function (config) {
     basePath: '',
     frameworks: ['jasmine'],
     files: [
-      'bower_components/jquery/jquery.js',
+      'bower_components/jquery/dist/jquery.js',
       'bower_components/sinonjs/sinon.js',
       'bower_components/jasmine-sinon/lib/jasmine-sinon.js',
       'bower_components/angular/angular.js',
@@ -14,6 +14,20 @@ module.exports = function (config) {
       'test/mocks/**/*.js',
       'test/spec/**/*.coffee'
     ],
+    preprocessors: {
+      '**/*.coffee': ['coffee']
+    },
+    coffeePreprocessor: {
+      // options passed to the coffee compiler
+      options: {
+        bare: true,
+        sourceMap: false
+      },
+      // transforming the filenames
+      transformPath: function (path) {
+        return path.replace(/\.coffee$/, '.js');
+      }
+    },
     exclude: [],
     reporters: ['dots'],
     autoWatch: false,